+++ src/stack.c
@@ -155,7 +155,7 @@ boolean safestack(void) {
 ** 'OPSTACKSIZE' numeric or string entries. It returns a pointer to
 ** the base of the stack
 */
-int32 *make_opstack(void) {
+size_t *make_opstack(void) {
   basicvars.stacktop.bytesp-=ALIGNSIZE(stack_opstack);
 #ifdef DEBUG
   if (basicvars.debug_flags.stack) fprintf(stderr, "stack.c:make_opstack: stacktop=%p, stacklimit=%p, OPSTACKSIZE*LARGEST_ENTRY=%lX\n", basicvars.stacktop.bytesp, basicvars.stacklimit.bytesp, (long unsigned int)OPSTACKSIZE*LARGEST_ENTRY);
